requirements.txt
alembic==1.4.3
Babel==2.9.1
blinker==1.5
certifi==2022.5.18
chardet==4.0.0
charset-normalizer==2.1.1
click==7.1.2
cmake==3.25.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==1.1.4
Flask-Admin==1.5.8
Flask-BabelEx==0.9.4
Flask-CKEditor==0.4.6
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.7.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==2.0.1
gunicorn==20.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.6
MarkupSafe==1.1.1
numpy==1.23.4
paho-mqtt==1.6.1
passlib==1.7.4
pika==1.3.1
Pillow==9.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-editor==1.0.4
pytz==2022.1
qrcode==7.3.1
requests==2.25.1
six==1.16.0
speaklater==1.3
SQLAlchemy==1.3.24
urllib3==1.26.9
Werkzeug==1.0.1
WTForms==2.2.1
PyJWT==2.8.0
flask-jwt-extended==4.3.0
pycryptodome==3.19.0
lxml==4.9.3
pypinyin==0.49.0

